Update version setup to use `BUILD_NUMBER`
By default, the release toolkit expect `VERSION_SHORT` to be in the
`3.2` (or `3.2.1` for hotfix) format and `VERSION_LONG`, `3.2.0.0` (or
`3.2.1.0` for hotfix).

This project uses a more convention (or old school?) build number
(`VERSION_LONG`) approach where the value is an increasing integer.

To account for this setup, we need to add a new version property,
`BUILD_NUMBER`.

Notice that the resulting setup is a bit unclear but that's an issue at
the release toolkit side.