Configure SwiftGen to generate nested enum, `L10n.Localizable`
This is preparatory work to feed more than one `.strings` file to SwiftGen.

By preemptively generating the nested enum, we ensure the future diff
where we'll add a new file to the configuration will be small and easier
to review. Moreover, while this diff is humongous, there is no logic
change and we can trust that if the app compiles the changes are
correct.