Update version details to 7.24.1
Something went wrong in our automation at some point and now we are
stuck with a build for version 7.24.0 with number 9, instead of the
expected 7.24.0.4. See
https://buildkite.com/automattic/pocket-casts-ios/builds/1535

The only option I see at the moment is to change build numbering
altogether, going back to an incrementing integer number.

Since our build number is already "compromised" I won't hurt to try an
Hail Mary: create a new version and hope that TestFlight will allow us
to use a different build number format. After all, that's what happened
when going from 7.20.2 (886) to 7.21 (7.21.0.0).