[Logging] Buffer FileLog to Prevent Frequent File Writes
This is an attempted fix for GitHub issue #17.
Specifically it appears that FileLog writes messages to the
filesystem too frequently, negatively impacting user battery.

The fix is to only write the log entries to the filesystem either when a
specific threshold is hit or when the user backgrounds the app.