Move creation of backmerge PR to the triggering lanes
Instead of creating the intermediate branches in the triggering lanes line `code_freeze`/`new_beta_release`/`finalize_release`/`finalize_hotfix_release` but only creating the backmerge PR during the `build_and_upload_app_store_connect` lane run on CI and only after the build was finished and uploaded to TestFlight.

That way the backmerge PR is created as soon as the build is triggered, instead of having to wait for the CI build to finish to be able to review and merge it. This also means that the CI checks on the backmerge PR itself will start sooner rather than only starting after the beta/final release build is done, allowing parallelization of those 2 CI builds.