Make `publish-release.yml` checkout release branch
Failing to checkout the release branch first would run this code on the detached HEAD commit that is prior to the version bump made by `new_beta_release` or `finalize_release`. As a consequence, it would
 - Create the GitHub Release with the wrong name, based on the version before the version bump
  - Create the git tag on the wrong commit

This commit also makes `checkout-release-branch.sh` fallback to derive the `RELEASE_VERSION` from the `BUILDKITE_BRANCH` if `RELEASE_VERSION` is not passed explicitly and the current branch name starts with `release/`. This is so that those changes can work even on scenarios that we have not yet updated to pass the `RELEASE_VERSION => '%version%` value explicitly during the corresponding `Task::buildkite`

See p1739836633853429/1739805357.834669-slack-C029BMLUGRX