Create a new methods for setting up the database
While using GRDB, we get some table locks if we execute "PRAGMA user_version" and then execute the migrations. I'm not sure why this happens but separating the calls make both to work without any issue.
Given that in the long term we should wrap reads in read {} blocks and write in write {} blocks this seems like a ok change for now.