Avoid replace action when changing a single episode
This _should_ only happen when the up next queue is empty but we theorize that outdated client data could lead one client to send a replace action and the other to send some other information. In theory, if we avoid the replace, we may have fewer issues with queues being cleared.