Reload URLSession cache to force Discover reload
This is only a development option so I think it's fine to be broader about our cache removal here.

I wonder why we have this double caching layer going on here. Why not use a separate URLSession with its own cache instead of reimplementing our own caching logic?

Alternatively, we could use a separate URLSession based on the shared one and clear that as necessary. This seems overkill for a dev option, though.