Update SwiftGen to latest version via CocoaPods
Aside from the fact that it's good to keep libraries up to date,
bumping SwiftGen from 6.5.1 to 6.6.0+ is a necessary prerequisite to
migrate from CocoaPods to SwiftPM for sourcing the library as the
SwiftGen SwiftPM plugin starts from version 6.6.0,
https://github.com/SwiftGen/SwiftGenPlugin/tags

Unfortunately for code review, the update introduces a breaking change.
As described in https://github.com/SwiftGen/SwiftGen/releases/tag/6.6.0,
"custom lookupFunctions will need to accept 3 arguments (table, key and
value), up from 2 arguments (table and key)."

This results in many changes in the generated code.