Limit playback position changes when switching episodes
When the changes position on the scrubber from the lock screen and we are switching episodes, it seems that we sometimes run into race conditions where the episode is skipped. The remote command doesn't give us any info about the media item but we can track when the episode changes from our own internal notifications. For now, I've set a delay of 2 seconds since that seems like a reasonable amount of time where trying again would work.