Fix PaidStoryWall analytics
This PR updates paywall presentation to use different logic for end of year 2025.
Instead of showing a Paywall view on front we use a story that is shown at that step.
This influences some of the logic around moving forward and backward on the plus story, because we are no longer overlaying the paywal on top of an existing story.
It also ensures that analytics events are sent only once when entering the paywall and not when preloading.